+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Example for `step_novel()` now better illustrates how it works. (@Edgar-Zamora, #1248)
 
+* `recipe()` now works with sparse tibbles. (#1364)
+
 # recipes 1.1.0
 
 ## Improvements